text/requirements.txt
Flask==2.3.2
markdown==3.4.4
pdfplumber==0.10.2
python-docx==0.8.11
beautifulsoup4==4.12.2
